pyproject.toml
[project]
name = "opinion-galaxy-mvp"
version = "0.1.0"
description = ""
authors = [{ name = "edegp", email = "a.yuhi1164@gmail.com" }]
readme = "README.md"
requires-python = "3.12.8"
dependencies = [
    "pandas (==2.2.3)",
    "numpy (==1.26.4)",
    "plotly (==5.24.1)",
    "pydantic (>=2.10.5,<3.0.0)",
    "pandera (>=0.22.1,<0.23.0)",
    "geopandas (>=1.0.1,<2.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)",
    "streamlit (>=1.41.1,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
]

[tool.poetry]
packages = [{ include = "src", from = "." }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
